requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.6
django-cors-headers==4.2.0
django-easy-audit==1.3.5
django-filter==23.3
djangorestframework==3.14.0
idna==3.4
psycopg2-binary==2.9.9
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
